+++ app/utils/jobScraper.ts
@@ -1,5 +1,5 @@
 import FirecrawlApp, { ScrapeResponse } from "@mendable/firecrawl-js";
-import { extractJobData, JobData } from "./aiEnhancer";
+import { extractJobData, JobData } from "../utils/jobDataExtractor";
 
 const app = new FirecrawlApp({
   apiKey: process.env.FIRECRAWL_API_KEY || "",
